Moved Header into components, removed unwanted files
