--- app/head.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-const Head = () => {
-    return (
-        <>
-            <title>DEV NEWS</title>
-            <meta name="viewport" content="width=device-width,initial-scale=1" />
-            <link rel="icon" href="/favicon.ico" />
-        </>
-    )
-}
-
-export default Head
